Crate rustc_codegen_llvm

The Rust compiler.

Note

This API is completely unstable and subject to change.

Modules

abi ğŸ”’

allocator ğŸ”’

asm ğŸ”’

attributes ğŸ”’

Set and unset common attributes on LLVM values.

back ğŸ”’

base ğŸ”’

Codegen the MIR to the LLVM IR.

builder ğŸ”’

callee ğŸ”’

Handles codegen of callees as well as other call-related things. Callees are a superset of normal rust values and sometimes have different representations. In particular, top-level fn items and methods are represented as just a fn ptr and not a full closure.

common ğŸ”’

Code that is useful in various codegen modules.

consts ğŸ”’

context ğŸ”’

coverageinfo ğŸ”’

debuginfo ğŸ”’

Debug Info Module

declare ğŸ”’

Declare various LLVM values.

errors ğŸ”’

intrinsic ğŸ”’

llvm

llvm_ ğŸ”’

llvm_util ğŸ”’

mono_item ğŸ”’

type_ ğŸ”’

type_of ğŸ”’

va_arg ğŸ”’

value ğŸ”’

Structs

LlvmCodegenBackend

ModuleLlvm

TimeTraceProfiler ğŸ”’

Functions

target_features

Used to generate cfg variables and apply features Must express features in the way Rust understands them