Module rustc_codegen_llvm::asm

Functions

a64_vreg_index 🔒

If the register is an AArch64 vector register then return its index.

dummy_output_type 🔒

Type to use for outputs that are discarded. It doesn’t really matter what the type is, as long as it is valid for the constraint code.

inline_asm_call 🔒

llvm_asm_scalar_type 🔒

Helper function to get the LLVM type for a Scalar. Pointers are returned as the equivalent integer type.

llvm_fixup_input 🔒

Fix up an input value to work around LLVM bugs.

llvm_fixup_output 🔒

Fix up an output value to work around LLVM bugs.

llvm_fixup_output_type 🔒

Output type to use for llvm_fixup_output.

modifier_to_llvm 🔒

Converts a modifier into LLVM’s equivalent modifier.

reg_to_llvm 🔒

Converts a register class to an LLVM constraint code.

xmm_reg_index 🔒

If the register is an xmm/ymm/zmm register then return its index.