Module rustc_codegen_llvm::attributes

Set and unset common attributes on LLVM values.

Enums

InlineAttr

InstructionSetAttr

OptimizeAttr

Functions

apply_to_callsite

apply_to_llfn

create_alloc_family_attr 🔒

default_optimisation_attrs 🔒

Get the default optimizations attrs for a function.

frame_pointer_type_attr

from_fn_attrs

Composite function which sets LLVM attributes for function depending on its AST (#[attribute]) attributes.

inline_attr 🔒

Get LLVM attribute for the provided inline heuristic.

instrument_function_attr 🔒

Tell LLVM what instrument function to insert.

non_lazy_bind_attr

Get the NonLazyBind LLVM attribute, if the codegen options allow skipping the PLT.

probestack_attr 🔒

sanitize_attrs

Get LLVM sanitize attributes.

stackprotector_attr 🔒

target_cpu_attr

tune_cpu_attr

uwtable_attr

Tell LLVM to emit or not emit the information necessary to unwind the stack for the function.

wasm_import_module 🔒