rustc_codegen_llvm::builder::autodiff

Function generate_enzyme_call

fn generate_enzyme_call<'ll>(
    cx: &GenericCx<'ll, SCx<'ll>>,
    fn_to_diff: &'ll Value,
    outer_fn: &'ll Value,
    attrs: AutoDiffAttrs,
)

When differentiating fn_to_diff, take a outer_fn and generate another function with expected naming and calling conventions1 which will be discovered by the enzyme LLVM pass and its body populated with the differentiated fn_to_diff. outer_fn is then modified to have a call to the generated function and handle the differences between the Rust calling convention and Enzyme.

https://enzyme.mit.edu/getting_started/CallingConvention/ ↩