Struct rustc_codegen_llvm::builder::Builder

pub struct Builder<'a, 'll, 'tcx> {
    pub llbuilder: &'ll mut Builder<'ll>,
    pub cx: &'a CodegenCx<'ll, 'tcx>,
}

Fields
llbuilder: &'ll mut Builder<'ll>cx: &'a CodegenCx<'ll, 'tcx>
Implementations

impl<'a, 'll, 'tcx> Builder<'a, 'll, 'tcx>

fn with_cx(cx: &'a CodegenCx<'ll, 'tcx>) -> Self

pub fn llfn(&self) -> &'ll Value

fn position_at_start(&mut self, llbb: &'ll BasicBlock)

fn align_metadata(&mut self, load: &'ll Value, align: Align)

fn noundef_metadata(&mut self, load: &'ll Value)

pub fn minnum(&mut self, lhs: &'ll Value, rhs: &'ll Value) -> &'ll Value

pub fn maxnum(&mut self, lhs: &'ll Value, rhs: &'ll Value) -> &'ll Value

pub fn insert_element( &mut self, vec: &'ll Value, elt: &'ll Value, idx: &'ll Value ) -> &'ll Value

pub fn shuffle_vector( &mut self, v1: &'ll Value, v2: &'ll Value, mask: &'ll Value ) -> &'ll Value

pub fn vector_reduce_fadd( &mut self, acc: &'ll Value, src: &'ll Value ) -> &'ll Value

pub fn vector_reduce_fmul( &mut self, acc: &'ll Value, src: &'ll Value ) -> &'ll Value

pub fn vector_reduce_fadd_fast( &mut self, acc: &'ll Value, src: &'ll Value ) -> &'ll Value

pub fn vector_reduce_fmul_fast( &mut self, acc: &'ll Value, src: &'ll Value ) -> &'ll Value

pub fn vector_reduce_add(&mut self, src: &'ll Value) -> &'ll Value

pub fn vector_reduce_mul(&mut self, src: &'ll Value) -> &'ll Value

pub fn vector_reduce_and(&mut self, src: &'ll Value) -> &'ll Value

pub fn vector_reduce_or(&mut self, src: &'ll Value) -> &'ll Value

pub fn vector_reduce_xor(&mut self, src: &'ll Value) -> &'ll Value

pub fn vector_reduce_fmin(&mut self, src: &'ll Value) -> &'ll Value

pub fn vector_reduce_fmax(&mut self, src: &'ll Value) -> &'ll Value

pub fn vector_reduce_fmin_fast(&mut self, src: &'ll Value) -> &'ll Value

pub fn vector_reduce_fmax_fast(&mut self, src: &'ll Value) -> &'ll Value

pub fn vector_reduce_min( &mut self, src: &'ll Value, is_signed: bool ) -> &'ll Value

pub fn vector_reduce_max( &mut self, src: &'ll Value, is_signed: bool ) -> &'ll Value

pub fn add_clause(&mut self, landing_pad: &'ll Value, clause: &'ll Value)

pub fn catch_ret( &mut self, funclet: &Funclet<'ll>, unwind: &'ll BasicBlock ) -> &'ll Value

fn check_call<'b>( &mut self, typ: &str, fn_ty: &'ll Type, llfn: &'ll Value, args: &'b [&'ll Value] ) -> Cow<'b, [&'ll Value]>

pub fn va_arg(&mut self, list: &'ll Value, ty: &'ll Type) -> &'ll Value

pub(crate) fn call_intrinsic( &mut self, intrinsic: &str, args: &[&'ll Value] ) -> &'ll Value

fn call_lifetime_intrinsic( &mut self, intrinsic: &str, ptr: &'ll Value, size: Size )

pub(crate) fn phi( &mut self, ty: &'ll Type, vals: &[&'ll Value], bbs: &[&'ll BasicBlock] ) -> &'ll Value

fn add_incoming_to_phi( &mut self, phi: &'ll Value, val: &'ll Value, bb: &'ll BasicBlock )

fn fptoint_sat( &mut self, signed: bool, val: &'ll Value, dest_ty: &'ll Type ) -> &'ll Value

pub(crate) fn landing_pad( &mut self, ty: &'ll Type, pers_fn: &'ll Value, num_clauses: usize ) -> &'ll Value

fn cfi_type_test( &mut self, fn_attrs: Option<&CodegenFnAttrs>, fn_abi: Option<&FnAbi<'tcx, Ty<'tcx>>>, llfn: &'ll Value )

fn kcfi_operand_bundle( &mut self, fn_attrs: Option<&CodegenFnAttrs>, fn_abi: Option<&FnAbi<'tcx, Ty<'tcx>>>, llfn: &'ll Value ) -> Option<OperandBundleDef<'ll>>

impl<'tcx> Builder<'_, '_, 'tcx>

fn set_function_source_hash( &mut self, instance: Instance<'tcx>, function_source_hash: u64 ) -> bool

Returns true if the function source hash was added to the coverage map (even if it had already been added, for this instance). Returns false only if -C instrument-coverage is not enabled (a coverage map is not being generated).

fn add_coverage_counter( &mut self, instance: Instance<'tcx>, id: CounterId, region: CodeRegion ) -> bool

Returns true if the counter was added to the coverage map; false if -C instrument-coverage is not enabled (a coverage map is not being generated).

fn add_coverage_counter_expression( &mut self, instance: Instance<'tcx>, id: ExpressionId, lhs: Operand, op: Op, rhs: Operand, region: Option<CodeRegion> ) -> bool

Returns true if the expression was added to the coverage map; false if -C instrument-coverage is not enabled (a coverage map is not being generated).

fn add_coverage_unreachable( &mut self, instance: Instance<'tcx>, region: CodeRegion ) -> bool

Returns true if the region was added to the coverage map; false if -C instrument-coverage is not enabled (a coverage map is not being generated).