rustc_codegen_llvm
Module callee

Handles codegen of callees as well as other call-related things. Callees are a superset of normal rust values and sometimes have different representations. In particular, top-level fn items and methods are represented as just a fn ptr and not a full closure.

Functions

get_fn 🔒

Codegens a reference to a fn/method item, monomorphizing and inlining as it goes.