rustc_codegen_llvm::context

Struct FullCx

pub(crate) struct FullCx<'ll, 'tcx> {

    pub tcx: TyCtxt<'tcx>,
    pub scx: GenericCx<'ll, SCx<'ll>>,
    pub use_dll_storage_attrs: bool,
    pub tls_model: ThreadLocalMode,
    pub codegen_unit: &'tcx CodegenUnit<'tcx>,
    pub instances: RefCell<FxHashMap<Instance<'tcx>, &'ll Value>>,
    pub vtables: RefCell<FxHashMap<(Ty<'tcx>, Option<ExistentialTraitRef<'tcx>>), &'ll Value>>,
    pub const_str_cache: RefCell<FxHashMap<String, &'ll Value>>,
    pub const_globals: RefCell<FxHashMap<&'ll Value, &'ll Value>>,
    pub statics_to_rauw: RefCell<Vec<(&'ll Value, &'ll Value)>>,
    pub used_statics: Vec<&'ll Value>,
    pub compiler_used_statics: Vec<&'ll Value>,
    pub type_lowering: RefCell<FxHashMap<(Ty<'tcx>, Option<VariantIdx>), &'ll Type>>,
    pub scalar_lltypes: RefCell<FxHashMap<Ty<'tcx>, &'ll Type>>,
    pub coverage_cx: Option<CguCoverageContext<'ll, 'tcx>>,
    pub dbg_cx: Option<CodegenUnitDebugContext<'ll, 'tcx>>,
    eh_personality: Cell<Option<&'ll Value>>,
    eh_catch_typeinfo: Cell<Option<&'ll Value>>,
    pub rust_try_fn: Cell<Option<(&'ll Type, &'ll Value)>>,
    intrinsics: RefCell<FxHashMap<(Cow<'static, str>, SmallVec<[&'ll Type; 2]>), (&'ll Type, &'ll Value)>>,
    local_gen_sym_counter: Cell<usize>,
    pub renamed_statics: RefCell<FxHashMap<DefId, &'ll Value>>,

}

Fields
tcx: TyCtxt<'tcx>scx: GenericCx<'ll, SCx<'ll>>use_dll_storage_attrs: booltls_model: ThreadLocalModecodegen_unit: &'tcx CodegenUnit<'tcx>instances: RefCell<FxHashMap<Instance<'tcx>, &'ll Value>>

Cache instances of monomorphic and polymorphic items

vtables: RefCell<FxHashMap<(Ty<'tcx>, Option<ExistentialTraitRef<'tcx>>), &'ll Value>>

Cache generated vtables

const_str_cache: RefCell<FxHashMap<String, &'ll Value>>

Cache of constant strings,

const_globals: RefCell<FxHashMap<&'ll Value, &'ll Value>>

Cache of emitted const globals (value -> global)

statics_to_rauw: RefCell<Vec<(&'ll Value, &'ll Value)>>

List of globals for static variables which need to be passed to the LLVM function ReplaceAllUsesWith (RAUW) when codegen is complete. (We have to make sure we don’t invalidate any Values referring to constants.)

used_statics: Vec<&'ll Value>

Statics that will be placed in the llvm.used variable See https://llvm.org/docs/LangRef.html#the-llvm-used-global-variable for details

compiler_used_statics: Vec<&'ll Value>

Statics that will be placed in the llvm.compiler.used variable See https://llvm.org/docs/LangRef.html#the-llvm-compiler-used-global-variable for details

type_lowering: RefCell<FxHashMap<(Ty<'tcx>, Option<VariantIdx>), &'ll Type>>

Mapping of non-scalar types to llvm types.

scalar_lltypes: RefCell<FxHashMap<Ty<'tcx>, &'ll Type>>

Mapping of scalar types to llvm types.

coverage_cx: Option<CguCoverageContext<'ll, 'tcx>>

Extra per-CGU codegen state needed when coverage instrumentation is enabled.

dbg_cx: Option<CodegenUnitDebugContext<'ll, 'tcx>>eh_personality: Cell<Option<&'ll Value>>eh_catch_typeinfo: Cell<Option<&'ll Value>>rust_try_fn: Cell<Option<(&'ll Type, &'ll Value)>>intrinsics: RefCell<FxHashMap<(Cow<'static, str>, SmallVec<[&'ll Type; 2]>), (&'ll Type, &'ll Value)>>local_gen_sym_counter: Cell<usize>

A counter that is used for generating local symbol names

renamed_statics: RefCell<FxHashMap<DefId, &'ll Value>>

codegen_static will sometimes create a second global variable with a different type and clear the symbol name of the original global. global_asm! needs to be able to find this new global so that it can compute the correct mangled symbol name to insert into the asm.