rustc_codegen_llvm::context

Struct GenericCx

pub(crate) struct GenericCx<'ll, T: Borrow<SCx<'ll>>>(T, PhantomData<SCx<'ll>>);

Tuple Fields
0: T1: PhantomData<SCx<'ll>>
Implementations

impl<'ll, CX: Borrow<SCx<'ll>>> GenericCx<'ll, CX>

pub(crate) fn const_array( &self, ty: &'ll Type, elts: &[&'ll Value], ) -> &'ll Value

pub(crate) fn const_bytes(&self, bytes: &[u8]) -> &'ll Value

pub(crate) fn const_get_elt(&self, v: &'ll Value, idx: u64) -> &'ll Value

pub(crate) fn const_null(&self, t: &'ll Type) -> &'ll Value

impl<'ll> GenericCx<'ll, FullCx<'ll, '_>>

pub(crate) fn const_bitcast(&self, val: &'ll Value, ty: &'ll Type) -> &'ll Value

pub(crate) fn const_pointercast( &self, val: &'ll Value, ty: &'ll Type, ) -> &'ll Value

pub(crate) fn static_addr_of_mut( &self, cv: &'ll Value, align: Align, kind: Option<&str>, ) -> &'ll Value

Create a global variable.

The returned global variable is a pointer in the default address space for globals. Fails if a symbol with the given name already exists.

pub(crate) fn static_addr_of_impl( &self, cv: &'ll Value, align: Align, kind: Option<&str>, ) -> &'ll Value

Create a global constant.

The returned global variable is a pointer in the default address space for globals.

pub(crate) fn get_static(&self, def_id: DefId) -> &'ll Value

fn get_static_inner(&self, def_id: DefId, llty: &'ll Type) -> &'ll Value

fn codegen_static_item(&mut self, def_id: DefId)

pub(crate) fn add_used_global(&mut self, global: &'ll Value)

Add a global value to a list to be stored in the llvm.used variable, an array of ptr.

pub(crate) fn add_compiler_used_global(&mut self, global: &'ll Value)

Add a global value to a list to be stored in the llvm.compiler.used variable, an array of ptr.

impl<'ll, 'tcx> GenericCx<'ll, FullCx<'ll, 'tcx>>

pub(crate) fn new( tcx: TyCtxt<'tcx>, codegen_unit: &'tcx CodegenUnit<'tcx>, llvm_module: &'ll ModuleLlvm, ) -> Self

pub(crate) fn statics_to_rauw(&self) -> &RefCell<Vec<(&'ll Value, &'ll Value)>>

pub(crate) fn coverage_cx(&self) -> &CguCoverageContext<'ll, 'tcx>

Extra state that is only available when coverage instrumentation is enabled.

pub(crate) fn create_used_variable_impl( &self, name: &'static CStr, values: &[&'ll Value], )

impl<'ll> GenericCx<'ll, SCx<'ll>>

pub(crate) fn get_return_type(&self, ty: &'ll Type) -> &'ll Type

pub(crate) fn get_type_of_global(&self, val: &'ll Value) -> &'ll Type

pub(crate) fn val_ty(&self, v: &'ll Value) -> &'ll Type

impl<'ll> GenericCx<'ll, SCx<'ll>>

pub(crate) fn new( llmod: &'ll Module, llcx: &'ll Context, pointer_size: Size, ) -> Self

impl<'ll, CX: Borrow<SCx<'ll>>> GenericCx<'ll, CX>

pub(crate) fn get_metadata_value(&self, metadata: &'ll Metadata) -> &'ll Value

pub(crate) fn get_const_int(&self, ty: &'ll Type, val: u64) -> &'ll Value

pub(crate) fn get_const_i64(&self, n: u64) -> &'ll Value

pub(crate) fn get_const_i32(&self, n: u64) -> &'ll Value

pub(crate) fn get_const_i16(&self, n: u64) -> &'ll Value

pub(crate) fn get_const_i8(&self, n: u64) -> &'ll Value

pub(crate) fn get_function(&self, name: &str) -> Option<&'ll Value>

pub(crate) fn get_md_kind_id(&self, name: &str) -> MetadataKindId

pub(crate) fn create_metadata(&self, name: &[u8]) -> &'ll Metadata

pub(crate) fn get_functions(&self) -> Vec<&'ll Value>

impl<'ll> GenericCx<'ll, FullCx<'ll, '_>>

pub(crate) fn get_intrinsic( &self, base_name: Cow<'static, str>, type_params: &[&'ll Type], ) -> (&'ll Type, &'ll Value)

fn declare_intrinsic( &self, base_name: &str, type_params: &[&'ll Type], ) -> (&'ll Type, &'ll Value)

pub(crate) fn eh_catch_typeinfo(&self) -> &'ll Value

impl GenericCx<'_, FullCx<'_, '_>>

pub(crate) fn generate_local_symbol_name(&self, prefix: &str) -> String

Generates a new symbol name with the given prefix. This symbol name must only be used for definitions with internal or private linkage.

impl<'ll, CX: Borrow<SCx<'ll>>> GenericCx<'ll, CX>

pub(crate) fn set_metadata<'a>( &self, val: &'a Value, kind_id: impl Into<MetadataKindId>, md: &'ll Metadata, )

A wrapper for llvm::LLVMSetMetadata, but it takes Metadata as a parameter instead of Value.

impl<'ll, 'tcx> GenericCx<'ll, FullCx<'ll, 'tcx>>

pub(crate) fn coverageinfo_finalize(&mut self)

fn covfun_section_name(&self) -> &CStr

Returns the section name to use when embedding per-function coverage information in the object file, according to the target’s object file format. LLVM’s coverage tools use information from this section when producing coverage reports.

Typical values are:

__llvm_covfun on Linux

__LLVM_COV,__llvm_covfun on macOS (includes __LLVM_COV, segment prefix)

.lcovfun$M on Windows (includes $M sorting suffix)

fn ensure_pgo_func_name_var(&self, instance: Instance<'tcx>) -> &'ll Value

For LLVM codegen, returns a function-specific Value for a global string, to hold the function name passed to LLVM intrinsic instrprof.increment(). The Value is only created once per instance. Multiple invocations with the same instance return the same Value.

This has the side-effect of causing coverage codegen to consider this function “used”, making it eligible to emit an associated covfun record.

impl<'ll> GenericCx<'ll, FullCx<'ll, '_>>

fn lookup_debug_loc(&self, pos: BytePos) -> DebugLoc

Looks up debug source information about a BytePos.

fn create_template_type_parameter( &self, name: &str, actual_type_metadata: &'ll Metadata, ) -> &'ll Metadata

impl<'ll, CX: Borrow<SCx<'ll>>> GenericCx<'ll, CX>

pub(crate) fn declare_global(&self, name: &str, ty: &'ll Type) -> &'ll Value

Declare a global value.

If there’s a value with the same name already declared, the function will return its Value instead.

impl<'ll, 'tcx> GenericCx<'ll, FullCx<'ll, 'tcx>>

pub(crate) fn declare_cfn( &self, name: &str, unnamed: UnnamedAddr, fn_type: &'ll Type, ) -> &'ll Value

Declare a C ABI function.

Only use this for foreign function ABIs and glue. For Rust functions use declare_fn instead.

If there’s a value with the same name already declared, the function will update the declaration and return existing Value instead.

pub(crate) fn declare_entry_fn( &self, name: &str, callconv: CallConv, unnamed: UnnamedAddr, fn_type: &'ll Type, ) -> &'ll Value

Declare an entry Function

The ABI of this function can change depending on the target (although for now the same as declare_cfn)

If there’s a value with the same name already declared, the function will update the declaration and return existing Value instead.

pub(crate) fn declare_fn( &self, name: &str, fn_abi: &FnAbi<'tcx, Ty<'tcx>>, instance: Option<Instance<'tcx>>, ) -> &'ll Value

Declare a Rust function.

If there’s a value with the same name already declared, the function will update the declaration and return existing Value instead.

impl<'ll, CX: Borrow<SCx<'ll>>> GenericCx<'ll, CX>

pub(crate) fn define_global( &self, name: &str, ty: &'ll Type, ) -> Option<&'ll Value>

Declare a global with an intention to define it.

Use this function when you intend to define a global. This function will return None if the name already has a definition associated with it. In that case an error should be reported to the user, because it usually happens due to user’s fault (e.g., misuse of #[no_mangle] or #[export_name] attributes).

pub(crate) fn define_private_global(&self, ty: &'ll Type) -> &'ll Value

Declare a private global

Use this function when you intend to define a global without a name.

pub(crate) fn get_declared_value(&self, name: &str) -> Option<&'ll Value>

Gets declared value by name.

pub(crate) fn get_defined_value(&self, name: &str) -> Option<&'ll Value>

Gets defined or externally defined (AvailableExternally linkage) value by name.

impl GenericCx<'_, FullCx<'_, '_>>

pub(crate) fn assume_dso_local( &self, llval: &Value, is_declaration: bool, ) -> bool

Whether a definition or declaration can be assumed to be local to a group of libraries that form a single DSO or executable. Marks the local as DSO if so.

fn should_assume_dso_local(&self, llval: &Value, is_declaration: bool) -> bool

impl<'ll, CX: Borrow<SCx<'ll>>> GenericCx<'ll, CX>

pub(crate) fn type_named_struct(&self, name: &str) -> &'ll Type

pub(crate) fn set_struct_body( &self, ty: &'ll Type, els: &[&'ll Type], packed: bool, )

pub(crate) fn type_void(&self) -> &'ll Type

pub(crate) fn type_ix(&self, num_bits: u64) -> &'ll Type

x Creates an integer type with the given number of bits, e.g., i24

pub(crate) fn type_vector(&self, ty: &'ll Type, len: u64) -> &'ll Type

pub(crate) fn func_params_types(&self, ty: &'ll Type) -> Vec<&'ll Type>

impl<'ll, 'tcx> GenericCx<'ll, FullCx<'ll, 'tcx>>

pub(crate) fn type_bool(&self) -> &'ll Type

pub(crate) fn type_int_from_ty(&self, t: IntTy) -> &'ll Type

pub(crate) fn type_uint_from_ty(&self, t: UintTy) -> &'ll Type

pub(crate) fn type_float_from_ty(&self, t: FloatTy) -> &'ll Type

pub(crate) fn type_padding_filler(&self, size: Size, align: Align) -> &'ll Type

Return an LLVM type that has at most the required alignment, and exactly the required size, as a best-effort padding array.

impl<'ll, CX: Borrow<SCx<'ll>>> GenericCx<'ll, CX>

pub(crate) fn llcx(&self) -> &'ll Context

pub(crate) fn llmod(&self) -> &'ll Module

pub(crate) fn isize_ty(&self) -> &'ll Type

pub(crate) fn type_variadic_func( &self, args: &[&'ll Type], ret: &'ll Type, ) -> &'ll Type

pub(crate) fn type_i1(&self) -> &'ll Type

pub(crate) fn type_struct(&self, els: &[&'ll Type], packed: bool) -> &'ll Type

impl<'a, 'tcx> GenericCx<'a, FullCx<'a, 'tcx>>

pub(crate) fn align_of(&self, ty: Ty<'tcx>) -> Align

pub(crate) fn size_of(&self, ty: Ty<'tcx>) -> Size

pub(crate) fn size_and_align_of(&self, ty: Ty<'tcx>) -> (Size, Align)