Module rustc_codegen_llvm::coverageinfo

Modules

mapgen

Structs

CrateCoverageContext

A context object for maintaining all state needed by the coverageinfo module.

Constants

UNUSED_FUNCTION_COUNTER_ID 🔒

VAR_ALIGN_BYTES 🔒

Functions

add_unused_function_coverage 🔒

codegen_unused_fn_and_counter 🔒

create_pgo_func_name_var 🔒

Calls llvm::createPGOFuncNameVar() with the given function instance’s mangled function name. The LLVM API returns an llvm::GlobalVariable containing the function name, with the specific variable name and linkage required by LLVM InstrProf source-based coverage instrumentation. Use bx.get_pgo_func_name_var() to ensure the variable is only created once per Instance.

declare_unused_fn 🔒

hash_bytes 🔒

hash_str 🔒

mapping_version 🔒

save_cov_data_to_mod 🔒

save_func_record_to_mod 🔒

write_filenames_section_to_buffer 🔒

write_mapping_to_buffer 🔒