Function rustc_codegen_llvm::coverageinfo::create_pgo_func_name_var

fn create_pgo_func_name_var<'ll, 'tcx>(
    cx: &CodegenCx<'ll, 'tcx>,
    instance: Instance<'tcx>,
) -> &'ll Value

Calls llvm::createPGOFuncNameVar() with the given function instance’s mangled function name. The LLVM API returns an llvm::GlobalVariable containing the function name, with the specific variable name and linkage required by LLVM InstrProf source-based coverage instrumentation. Use bx.get_pgo_func_name_var() to ensure the variable is only created once per Instance.