rustc_codegen_llvm::coverageinfo
Module map_data

Structs

FunctionCoverage 🔒

FunctionCoverageCollector 🔒

Holds all of the coverage mapping data associated with a function instance, collected during traversal of Coverage statements in the function’s MIR.

ZeroExpressions 🔒

Set of expression IDs that are known to always evaluate to zero. Any mapping or expression operand that refers to these expressions can have that reference replaced with a constant zero value.

Functions

is_zero_term 🔒

Returns true if the given term is known to have a value of zero, taking into account knowledge of which counters are unused and which expressions are always zero.