Module rustc_codegen_llvm::debuginfo::metadata::enums

Modules

cpp_like ğŸ”’

native ğŸ”’

Enums

DiscrResult ğŸ”’

Functions

build_c_style_enum_di_node ğŸ”’

Build the debuginfo node for a C-style enum, i.e. an enum the variants of which have no fields.

build_enum_type_di_node ğŸ”’

build_enum_variant_struct_type_di_node ğŸ”’

Build the debuginfo node for the struct type describing a single variant of an enum.

build_enumeration_type_di_node ğŸ”’

Build a DW_TAG_enumeration_type debuginfo node, with the given base type and variants. This is a helper function and does not register anything in the type map by itself.

build_generator_di_node ğŸ”’

build_generator_variant_struct_type_di_node

Build the struct type for describing a single generator state. See build_generator_variant_struct_type_di_node.

compute_discriminant_value ğŸ”’

Returns the discriminant value corresponding to the variant index.

tag_base_type ğŸ”’

Extract the type with which we want to describe the tag of the given enum or generator.