Function rustc_codegen_llvm::debuginfo::metadata::enums::cpp_like::build_union_fields_for_direct_tag_enum_or_generator

fn build_union_fields_for_direct_tag_enum_or_generator<'ll, 'tcx>(
    cx: &CodegenCx<'ll, 'tcx>,
    enum_type_and_layout: TyAndLayout<'tcx>,
    enum_type_di_node: &'ll DIType,
    variant_field_infos: &[VariantFieldInfo<'ll>],
    discr_type_di_node: &'ll DIType,
    tag_base_type: Ty<'tcx>,
    tag_field: usize,
    untagged_variant_index: Option<VariantIdx>
) -> SmallVec<[&'ll DIType; 16]>

This is a helper function shared between enums and generators that makes sure fields have the expect names.