rustc_codegen_llvm::debuginfo::metadata::enums

Function build_enumeration_type_di_node

fn build_enumeration_type_di_node<'ll, 'tcx>(
    cx: &GenericCx<'ll, FullCx<'ll, 'tcx>>,
    type_name: &str,
    base_type: Ty<'tcx>,
    enumerators: impl Iterator<Item = (Cow<'tcx, str>, u128)>,
    def_id: Option<DefId>,
    containing_scope: &'ll Metadata,
) -> &'ll Metadata

Build a DW_TAG_enumeration_type debuginfo node, with the given base type and variants. This is a helper function and does not register anything in the type map by itself.

variants is an iterator of (discr-value, variant-name).