rustc_codegen_llvm
Module declare

Declare various LLVM values.

Prefer using functions and methods from this module rather than calling LLVM functions directly. These functions do some additional work to ensure we do the right thing given the preconceptions of codegen.

Some useful guidelines:

Use declare_* family of methods if you are declaring, but are not interested in defining the Value they return.

Use define_* family of methods when you might be defining the Value.

When in doubt, define.

Functions

declare_raw_fn 🔒

Declare a function.