Module rustc_codegen_llvm::llvm_::ffi::coverageinfo

Structs

CounterMappingRegion

This struct provides LLVM’s representation of a “CoverageMappingRegion”, encoded into the coverage map, in accordance with the LLVM Code Coverage Mapping Format. The struct composes fields representing the Counter type and value(s) (injected counter ID, or expression type and operands), the source file (an indirect index into a “filenames array”, encoded separately), and source location (start and end positions of the represented code region).

Enums

RegionKind

Aligns with llvm::coverage::CounterMappingRegion::RegionKind