rustc_codegen_llvm
Module llvm_util

Structs

LLVMFeature ğŸ”’

Enums

TargetFeatureFoldStrength ğŸ”’

Statics

INIT ğŸ”’

Functions

backend_feature_name ğŸ”’

Returns a feature name for the given +feature or -feature string.

configure_llvm ğŸ”’ âš

get_version ğŸ”’

global_llvm_features ğŸ”’

The list of LLVM features computed from CLI flags (-Ctarget-cpu, -Ctarget-feature, --target and similar).

handle_native ğŸ”’

init ğŸ”’

llvm_target_features ğŸ”’

print ğŸ”’

print_passes ğŸ”’

print_target_features ğŸ”’

print_version ğŸ”’

require_inited ğŸ”’

target_cpu ğŸ”’

target_features

Used to generate cfg variables and apply features Must express features in the way Rust understands them

time_trace_profiler_finish ğŸ”’

to_llvm_features ğŸ”’

tune_cpu ğŸ”’