rustc_codegen_llvm

Module attributes

Set and unset common attributes on LLVM values.

Functions

apply_to_callsite ğŸ”’

apply_to_llfn ğŸ”’

backchain_attr ğŸ”’

create_alloc_family_attr ğŸ”’

default_optimisation_attrs ğŸ”’

Get the default optimizations attrs for a function.

frame_pointer_type_attr ğŸ”’

function_return_attr ğŸ”’

has_attr ğŸ”’

has_string_attr ğŸ”’

inline_attr ğŸ”’

Get LLVM attribute for the provided inline heuristic.

instrument_function_attr ğŸ”’

Tell LLVM what instrument function to insert.

llfn_attrs_from_instance ğŸ”’

Helper for FnAbi::apply_attrs_llfn: Composite function which sets LLVM attributes for function depending on its AST (#[attribute]) attributes.

nojumptables_attr ğŸ”’

non_lazy_bind_attr ğŸ”’

Get the NonLazyBind LLVM attribute, if the codegen options allow skipping the PLT.

patchable_function_entry_attrs ğŸ”’

probestack_attr ğŸ”’

remove_from_llfn ğŸ”’

remove_string_attr_from_llfn ğŸ”’

sanitize_attrs ğŸ”’

Get LLVM sanitize attributes.

stackprotector_attr ğŸ”’

target_cpu_attr ğŸ”’

tune_cpu_attr ğŸ”’

uwtable_attr ğŸ”’

Tell LLVM to emit or not emit the information necessary to unwind the stack for the function.

wasm_import_module ğŸ”’