Module rustc_codegen_llvm::base

Codegen the MIR to the LLVM IR.

Hopefully useful general knowledge about codegen:

There’s no way to find out the Ty type of a Value. Doing so would be “trying to get the eggs out of an omelette” (credit: pcwalton). You can, instead, find out its llvm::Type by calling val_ty, but one llvm::Type corresponds to many Tys; for instance, tup(int, int, int) and rec(x=int, y=int, z=int) will have the same llvm::Type.

Structs

ValueIter

Functions

compile_codegen_unit

iter_globals

linkage_to_llvm

set_link_section

visibility_to_llvm