rustc_codegen_llvm
Module coverageinfo

Modules

ffi 🔒

map_data 🔒

mapgen 🔒

Structs

CrateCoverageContext 🔒

A context object for maintaining all state needed by the coverageinfo module.

Functions

create_pgo_func_name_var 🔒

Calls llvm::createPGOFuncNameVar() with the given function instance’s mangled function name. The LLVM API returns an llvm::GlobalVariable containing the function name, with the specific variable name and linkage required by LLVM InstrProf source-based coverage instrumentation. Use bx.get_pgo_func_name_var() to ensure the variable is only created once per Instance.

hash_bytes 🔒

mapping_version 🔒

write_filenames_section_to_buffer 🔒

write_mapping_to_buffer 🔒