rustc_codegen_llvm::coverageinfo::mapgen
Module covfun

For each function that was instrumented for coverage, we need to embed its corresponding coverage mapping metadata inside the __llvm_covfun1 linker section of the final binary.

On Windows the section name is .lcovfun. ↩

Structs

CovfunRecord 🔒

Intermediate coverage metadata for a single function, used to help build the final record that will be embedded in the __llvm_covfun section.

Functions

fill_region_tables 🔒

Populates the mapping region tables in the current function’s covfun record.

generate_covfun_record 🔒

Generates the contents of the covfun record for this function, which contains the function’s coverage mapping data. The record is then stored as a global variable in the __llvm_covfun section.

prepare_covfun_record 🔒