rustc_codegen_llvm::debuginfo::metadata
Module enums

Modules

cpp_like 🔒

native 🔒

Enums

DiscrResult 🔒

Functions

build_c_style_enum_di_node 🔒

Build the debuginfo node for a C-style enum, i.e. an enum the variants of which have no fields.

build_coroutine_di_node 🔒

build_coroutine_variant_struct_type_di_node 🔒

Build the struct type for describing a single coroutine state. See build_coroutine_variant_struct_type_di_node.

build_enum_type_di_node 🔒

build_enum_variant_struct_type_di_node 🔒

Build the debuginfo node for the struct type describing a single variant of an enum.

build_enumeration_type_di_node 🔒

Build a DW_TAG_enumeration_type debuginfo node, with the given base type and variants. This is a helper function and does not register anything in the type map by itself.

compute_discriminant_value 🔒

Returns the discriminant value corresponding to the variant index.