Module rustc_codegen_llvm::debuginfo::metadata::enums::native

Structs

VariantMemberInfo 🔒

Information needed for building a DW_TAG_variant:

Functions

build_coroutine_di_node 🔒

Build the debuginfo node for a coroutine environment. It looks the same as the debuginfo for an enum. See build_enum_type_di_node for more information.

build_discr_member_di_node 🔒

Builds the DW_TAG_member describing where we can find the tag of an enum. Returns None if the enum does not have a tag.

build_enum_type_di_node 🔒

Build the debuginfo node for an enum type. The listing below shows how such a type looks like at the LLVM IR/DWARF level. It is a DW_TAG_structure_type with a single DW_TAG_variant_part that in turn contains a DW_TAG_variant for each variant of the enum. The variant-part also contains a single member describing the discriminant, and a nested struct type for each of the variants.

build_enum_variant_member_di_node 🔒

Build the debuginfo node for DW_TAG_variant:

build_enum_variant_part_di_node 🔒

Builds the DW_TAG_variant_part of an enum or coroutine debuginfo node: