Module rustc_codegen_llvm::llvm_util

Structs

LLVMFeature

Enums

TargetFeatureFoldStrength

Statics

INIT ğŸ”’

Functions

backend_feature_name ğŸ”’

Returns a feature name for the given +feature or -feature string.

check_tied_features

Given a map from target_features to whether they are enabled or disabled, ensure only valid combinations are allowed.

configure_llvm ğŸ”’ âš

get_version

global_llvm_features ğŸ”’

The list of LLVM features computed from CLI flags (-Ctarget-cpu, -Ctarget-feature, --target and similar).

handle_native ğŸ”’

init ğŸ”’

llvm_target_features ğŸ”’

print ğŸ”’

print_passes

print_target_features ğŸ”’

print_version

require_inited ğŸ”’

target_cpu

target_features

Used to generate cfg variables and apply features Must express features in the way Rust understands them

time_trace_profiler_finish

to_llvm_features

tune_cpu

Opaque Types

ğŸ”’